system-linux: re-apply ethtool on phy attachment
ethtool settings applied before the interface is in IFF_UP state
are going to be lost when phylink is being used.
This is the case with many modern NICs as well as when using SFP
modules, resulting in the speed, duplex, *pause and autoneg
settings not having any effect.

Reapply ethtool settings once the PHY is attached.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>