interface-ip: fix routes expiry time
During the testing process for the IOL INTACT cetification for IPv6, we discovered some issues:
'Test v6LC.2.2.14 (A): Router Lifetime Updated with Same Lifetime' Failed because Routes
are not removed at the correct expiry time. lifetime is set correctly but the condition to
remove the route is wrong

Description:
- TR1 transmits the Router Advertisement (Router Lifetime: 20 seconds).
- TN2 transmits a global Echo Request to the HUT every second for 19 seconds.
- The HUT should respond to the Echo Requests from TN2 using TR1 as a first hop.
- TR1 transmits the Router Advertisement (Router Lifetime: 20 seconds).
- TN2 transmits a global Echo Request to the HUT every second for 21 seconds.
- The HUT should respond to the Echo Requests from TN2 using TR1 as a first hop
until the Router Lifetime expires. In response to the final Echo Request, the HUT

Solution:
- The function interface_ip_valid_until_handler is called each 1 second to remove expired element the
route list. the condition to remove the route should be valid_until inferior or equal from time now.
Delay of one Second cause the test to fail.
- The precision of time is not efficient because we use only seconds from 'system_get_rtime'
- change the timer of interface_ip_valid_until_handler to 500 ms instead of 1s.

Signed-off-by: kaies Chaouch <kc@simonwunderlich.de>