interface-ip: overwrite resolv.conf file instead of renaming
Currently, the change tracking of resolv.conf works as follows:

1. Create a temporary resolv.conf with current entries.
2. Calculate CRC32 of both the temporary and the existing (old)
   resolv.conf file.
3. If the CRC checksum differs, the temporary file is renamed to
   replace the old resolv.conf.

However, this causes issues with ujail environments where the jail
has bind-mounted the resolv.conf from the host. In this case, the new
file would not be visible inside the jail until the jail was restarted.
This is particularly troublesome for jailed ntpd deamon running on
a device without a local DNS server, relying on a DNS info supplied
from DHCP (e.g. dumb AP). Very often, the DHCP lease comes later than
the ntpd is started, causing time synchronization to fail until
the process is manually restarted.

The proposed solution modifies the step 3 in the resolv.conf change
tracking above:

3. If the CRC checksum differs, the existing resolv.conf is overwritten
   in place with the new (current) entries (ones from the temporary
   file).

This approach, however, has a drawback - rename is atomic while
overwrite is not. To mitigate this issue, a file locking mechanism is
added (`flock`).

Fixes:
- openwrt/openwrt#10389
- openwrt/openwrt#10843

Signed-off-by: Dávid Benko <davidbenko@davidbenko.dev>